% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HawcClient.R
\name{HawcClient}
\alias{HawcClient}
\title{HawcClient}
\usage{
HawcClient(baseUrl)
}
\description{
This function sets up a HawcClient class, which contains functions
for specific get and post requests from HAWC APIs
}
